+++ project/src/app/services/login.services.spec.ts
@@ -1,71 +1,16 @@
-import { Injectable } from '@angular/core';
-import { AngularFire, AuthProviders, AuthMethods, FirebaseAuthState } from 'angularfire2';
-import * as firebase from 'firebase';
+/* tslint:disable:no-unused-variable */
 
-import 'rxjs/add/operator/toPromise';
+import { TestBed, async, inject } from '@angular/core/testing';
+import { LoginService } from './login.service';
 
-@Injectable()
-export class LoginService {
+describe('Service: Login', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [LoginService]
+    });
+  });
 
-  public isAuthenticated = false;
-  public displayName: string = '';
-  public photoUrl: string = '';
-
-  constructor(private af: AngularFire) {
-    console.log("Sparking a new LoginService");
-  }
-
-
-  private storeAuthInfo(authState: FirebaseAuthState): FirebaseAuthState {
-    if (authState) {
-      this.displayName = authState.auth.displayName;
-      this.photoUrl = authState.auth.photoURL;
-      this.isAuthenticated = true;
-      if (authState.google) {
-        localStorage.setItem('idToken', (authState.google as any).idToken);
-        localStorage.setItem('accessToken', (authState.google as any).accessToken);
-      }
-    }
-    return authState;
-  }
-
-  login(): firebase.Promise<FirebaseAuthState> {
-
-    const idToken = localStorage.getItem('idToken');
-    const accessToken = localStorage.getItem('accessToken');
-
-    if (idToken && accessToken) {
-
-      const authConfig = {
-        method: AuthMethods.OAuthToken,
-        provider: AuthProviders.Google
-      };
-      const credential = firebase.auth.GoogleAuthProvider.credential(idToken, accessToken);
-      return this.af.auth.login(credential, authConfig).then((authState) => {
-        console.log("Successful Token-based Login");
-        return this.storeAuthInfo(authState);
-      }).catch((err) => {
-        console.log("Error with auth token: " + err, " Clearing cached token..");
-        localStorage.setItem('idToken', '');
-        localStorage.setItem('accessToken', '');
-      });
-    } else {
-      // fall through to popup auth
-      return this.af.auth.login({
-        method: AuthMethods.Popup
-      }).then((authState) => {
-        console.log("Successful OAuth-based Login");
-        return this.storeAuthInfo(authState);
-      }).catch((err) => {
-        console.log(err);
-      });
-    }
-  }
-
-  logout() {
-    this.isAuthenticated = false;
-    this.displayName = this.photoUrl = '';
-    this.af.auth.logout();
-  }
-
-}
+  it('should ...', inject([LoginService], (service: LoginService) => {
+    expect(service).toBeTruthy();
+  }));
+});
